move cpuid_flags.h header to common
