+++ CMakeLists.txt
@@ -569,7 +569,7 @@ set (hs_exec_common_SRCS
     src/alloc.c
     src/scratch.c
     src/util/arch/x86/cpuid_flags.c
-    src/util/arch/x86/cpuid_flags.h
+    src/util/arch/common/cpuid_flags.h
     src/util/multibit.c
     )
